Remove unnecessary callmodel_p.h header
